Fix cleanup in tearDownClass method and delete container if it exists

+++ tests/test_app.py
@@ -53,14 +53,15 @@ class APITestCase(unittest.TestCase):
         connection_manager.rollback()
         connection_manager.put()
 
-    
+    @classmethod
     def tearDownClass(self):
         # Delete the content of the storage account
         FERTISCAN_STORAGE_URL = os.getenv("FERTISCAN_STORAGE_URL")
         with ContainerClient.from_connection_string(
             FERTISCAN_STORAGE_URL, container_name=f"user-{self.username}"
         ) as container_client:
-            container_client.delete_container()
+            if container_client.exists():
+                container_client.delete_container()
     
     def test_health(self):
         response = self.client.get("/health", headers=self.headers)
